v3.7.4 commit 8: pre-merge review fixes
Three improvements found by the v3.7.4 pre-merge code review:

1. child_agent_react_timeout_short|continuous typed as float
   (was int) to match the parent Config.react_timeout: float.
   YAML configurations writing 60.5 no longer silently coerce
   to 60. Pure type/default change, behavior unchanged.

2. JobHandle.poll() now passes join_timeout=0 when cancel()
   is triggered by an internal timeout transition. Pre-fix,
   any caller polling job state (JobService.has_active_jobs,
   the cancel_job tool's own poll() precheck) could block for
   up to job_cancel_join_timeout seconds. Polling stays
   non-blocking; explicit cancel/shutdown paths still get the
   bounded join via the configured timeout.

3. JobService.shutdown is now two-pass: pass 1 signal-cancels
   every running handle (no join), pass 2 joins each worker
   with the configured timeout. Total shutdown latency is
   bounded by cancel_join_timeout instead of the prior
   max_active * cancel_join_timeout (default 50s with 10
   wedged jobs). Daemon-fallback semantics unchanged.

cancel(join_timeout=0) is now a documented signal-only mode;
the join + warning are gated on `timeout > 0`.

+++ llamagent/core/config.py
@@ -145,8 +145,8 @@ _YAML_MAP = [
     (("child_agent", "max_plan_adjustments"), "child_agent_max_plan_adjustments", int),
     (("child_agent", "max_react_steps_short"), "child_agent_max_react_steps_short", int),
     (("child_agent", "max_react_steps_continuous"), "child_agent_max_react_steps_continuous", int),
-    (("child_agent", "react_timeout_short"), "child_agent_react_timeout_short", int),
-    (("child_agent", "react_timeout_continuous"), "child_agent_react_timeout_continuous", int),
+    (("child_agent", "react_timeout_short"), "child_agent_react_timeout_short", float),
+    (("child_agent", "react_timeout_continuous"), "child_agent_react_timeout_continuous", float),
     (("child_agent", "context_window_short"), "child_agent_context_window_short", int),
     (("child_agent", "context_window_continuous"), "child_agent_context_window_continuous", int),
     # v3.5
@@ -382,8 +382,8 @@ class Config:
         self.child_agent_max_plan_adjustments: int = 3
         self.child_agent_max_react_steps_short: int = 5
         self.child_agent_max_react_steps_continuous: int = 10
-        self.child_agent_react_timeout_short: int = 60
-        self.child_agent_react_timeout_continuous: int = 600
+        self.child_agent_react_timeout_short: float = 60.0
+        self.child_agent_react_timeout_continuous: float = 600.0
         self.child_agent_context_window_short: int = 10
         self.child_agent_context_window_continuous: int = 20
